// src/types/data.ts
export interface SourceData {
    id: string;
    intermediaryFirmName: string;
    intFirmClientNumber: string;
    initiatingFirmName: string;
    initiatingFirmIndustryID: string;
    addressLine1: string;
    city: string;
    country: string;
    postalCode: string;
    stateRegion: string;
  }